# apps/lms-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  service.AnswerSubmission:
    properties:
      question_id:
        type: string
      selected_option_id:
        type: string
      written_answer:
        type: string
    required:
    - question_id
    type: object
  service.AttemptAnswerResponse:
    properties:
      id:
        type: string
      is_correct:
        type: boolean
      marks_obtained:
        type: integer
      question_id:
        type: string
      selected_option_id:
        type: string
      written_answer:
        type: string
    type: object
  service.AttemptListResponse:
    properties:
      id:
        type: string
      obtained_marks:
        type: integer
      percentage:
        type: number
      started_at:
        type: string
      status:
        type: string
      student_id:
        type: string
      submitted_at:
        type: string
      test_series_id:
        type: string
      test_title:
        type: string
    type: object
  service.AttemptResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/service.AttemptAnswerResponse'
        type: array
      id:
        type: string
      obtained_marks:
        type: integer
      percentage:
        type: number
      started_at:
        type: string
      status:
        type: string
      student:
        $ref: '#/definitions/service.UserResponse'
      student_id:
        type: string
      submitted_at:
        type: string
      test_series:
        $ref: '#/definitions/service.TestSeriesResponse'
      test_series_id:
        type: string
      time_spent_seconds:
        type: integer
      total_marks:
        type: integer
    type: object
...
  service.StartAttemptRequest:
    properties:
      student_id:
        type: string
      test_series_id:
        type: string
    required:
    - student_id
    - test_series_id
    type: object
  service.SubmitAttemptRequest:
    properties:
      answers:
        items:
          $ref: '#/definitions/service.AnswerSubmission'
        type: array
      attempt_id:
        type: string
    required:
    - answers
    - attempt_id
    type: object
...
  service.TestSeriesResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      is_published:
        type: boolean
      passing_percentage:
        type: number
      show_correct_answers:
        type: boolean
      shuffle_questions:
        type: boolean
      slug:
        type: string
      title:
        type: string
      total_questions:
        type: integer
      updated_at:
        type: string
    type: object
...
  service.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login:
        type: string
      last_name:
        type: string
      phone:
        type: string
      profile_picture_url:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Learning Management System Backend API with Course, Test, and Student
    Management
  license:
    name: MIT
  title: LMS Backend API
  version: "1.0"
paths:
  /attempts:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of test attempts for authenticated student
      parameters:
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of attempts per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.AttemptListResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List student attempts
      tags:
      - Attempts
  /attempts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific test attempt
      parameters:
      - description: Attempt UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AttemptResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get attempt details
      tags:
      - Attempts
  /attempts/{id}/evaluate:
    post:
      consumes:
      - application/json
      description: Manually evaluate and score an attempt (admin only)
      parameters:
      - description: Attempt UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AttemptResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Evaluate an attempt
      tags:
      - Attempts
  /attempts/{id}/results:
    get:
      consumes:
      - application/json
      description: Retrieve detailed results and evaluation of a completed attempt
      parameters:
      - description: Attempt UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AttemptResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get attempt results
      tags:
      - Attempts
  /attempts/{id}/score:
    get:
      consumes:
      - application/json
      description: Retrieve score and percentage for a completed attempt
      parameters:
      - description: Attempt UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get attempt score
      tags:
      - Attempts
  /attempts/history/{test_series_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all attempts by user for a specific test
      parameters:
      - description: Test Series UUID
        in: path
        name: test_series_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.AttemptListResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get user attempt history
      tags:
      - Attempts
  /attempts/start:
    post:
      consumes:
      - application/json
      description: Begin a new test attempt for authenticated student
      parameters:
      - description: Start attempt request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.StartAttemptRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AttemptResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Start a new test attempt
      tags:
      - Attempts
  /attempts/submit:
    post:
      consumes:
      - application/json
      description: Submit answers for a test attempt
      parameters:
      - description: Submit attempt request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.SubmitAttemptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.AttemptResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Submit test answers
      tags:
      - Attempts
  /attempts/test/{test_series_id}:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all attempts for a test series (admin/instructor
        only)
      parameters:
      - description: Test Series UUID
        in: path
        name: test_series_id
        required: true
        type: string
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of attempts per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.AttemptListResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List all attempts for a test
      tags:
      - Attempts
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Add a JWT token to the header with ** Bearer &lt;JWT&gt; ** token
      to authorize"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"